+++ lib/src/televerse/api/raw_api.dart
@@ -4229,12 +4229,15 @@ class RawAPI {
   ///   “bold”, “italic”, “underline”, “strikethrough”, “spoiler”, and “custom_emoji” are ignored.
   /// - [textEntities] (List<MessageEntity>, optional): A JSON-serialized list of special entities
   ///   that appear in the gift text. Can be specified instead of `textParseMode`.
+  /// - [payForUpgrade] - Pass True to pay for the gift upgrade from the bot's
+  /// balance, thereby making the upgrade free for the receiver
   Future<bool> sendGift({
     required int userId,
     required String giftId,
     String? text,
     ParseMode? textParseMode,
     List<MessageEntity>? textEntities,
+    bool? payForUpgrade,
   }) async {
     final params = {
       "user_id": userId,
@@ -4242,11 +4245,108 @@ class RawAPI {
       "text": text,
       "text_parse_mode": textParseMode?.value,
       "text_entities": textEntities?.map((e) => e.toJson()).toList(),
+      "pay_for_upgrade": payForUpgrade,
     };
 
     return await _makeApiBoolCall(
       APIMethod.sendGift,
       payload: Payload(params),
     );
+  }
+
+  /// Verifies a user on behalf of the organization which is represented by
+  /// the bot. Returns True on success.
+  ///
+  /// Parameters:
+  /// - [userId] (int, required): Unique identifier of the target user.
+  /// - [customDescription] (String, optional): Custom description for
+  /// the verification; 0-70 characters. Must be empty if the organization
+  /// isn't allowed to provide a custom verification description.
+  Future<bool> verifyUser({
+    required int userId,
+    String? customDescription,
+  }) async {
+    final params = {
+      "user_id": userId,
+      "custom_description": customDescription,
+    };
+
+    final response = await _makeApiBoolCall(
+      APIMethod.verifyUser,
+      payload: Payload(params),
+    );
+
+    return response;
+  }
+
+  /// Verifies a chat on behalf of the organization which is represented by
+  /// the bot. Returns True on success.
+  ///
+  /// Parameters:
+  /// - [chatId] (ID, required): Unique identifier for the target chat or
+  /// username of the target channel (in the format @channelusername)
+  /// - [customDescription] (String, optional): Custom description for
+  /// the verification; 0-70 characters. Must be empty if the organization
+  /// isn't allowed to provide a custom verification description.
+  Future<bool> verifyChat({
+    required ID chatId,
+    String? customDescription,
+  }) async {
+    final params = {
+      "chat_id": chatId.id,
+      "custom_description": customDescription,
+    };
+
+    final response = await _makeApiBoolCall(
+      APIMethod.verifyChat,
+      payload: Payload(params),
+    );
+
+    return response;
+  }
+
+  /// Removes verification from a user who is currently verified on behalf
+  /// of the organization represented by the bot.
+  ///
+  /// Returns True on success.
+  ///
+  /// Parameters:
+  /// - [userId] - Unique identifier of the target user
+  Future<bool> removeUserVerification({
+    required int userId,
+  }) async {
+    final params = {
+      "user_id": userId,
+    };
+
+    final response = await _makeApiBoolCall(
+      APIMethod.removeUserVerification,
+      payload: Payload(params),
+    );
+
+    return response;
+  }
+
+  /// Removes verification from a chat that is currently verified on behalf of
+  /// the organization represented by the bot. Returns True on success.
+  ///
+  /// Returns True on success.
+  ///
+  /// Parameters:
+  /// - [chatId] - Unique identifier for the target chat or username of the
+  /// target channel (in the format @channelusername)
+  Future<bool> removeChatVerification({
+    required ID chatId,
+  }) async {
+    final params = {
+      "chat_id": chatId.id,
+    };
+
+    final response = await _makeApiBoolCall(
+      APIMethod.removeChatVerification,
+      payload: Payload(params),
+    );
+
+    return response;
   }
 }